S3 upload, bug fixes

+++ src/org/freeeed/main/ActionStaging.java
@@ -1,7 +1,3 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 package org.freeeed.main;
 
 import com.google.common.io.Files;
@@ -36,7 +32,10 @@ public class ActionStaging implements Runnable {
         History.appendToHistory("Project: " + processingParameters.getString(ParameterProcessing.PROJECT_NAME));
         // TODO better setting of dirs?
         String stagingDir = PackageArchive.stagingDir;
-        Files.deleteRecursively(new File(stagingDir));
+        File stagingDirFile = new File(stagingDir);
+        if (stagingDirFile.exists()) {
+            Files.deleteRecursively(new File(stagingDir));
+        }
         new File(stagingDir).mkdirs();
 
         String[] dirs = processingParameters.getStringArray(ParameterProcessing.PROJECT_INPUTS);
@@ -66,7 +65,10 @@ public class ActionStaging implements Runnable {
 
     private boolean downloadUri(String[] dirs) throws Exception {
         boolean anyDownload = false;
-        Files.deleteRecursively(new File(ParameterProcessing.DOWNLOAD_DIR));
+        File downloadDirFile = new File(ParameterProcessing.DOWNLOAD_DIR);
+        if (downloadDirFile.exists()) {
+            Files.deleteRecursively(downloadDirFile);
+        }
         new File(ParameterProcessing.DOWNLOAD_DIR).mkdirs();
 
         for (String dir : dirs) {
